+++ src/model_io/urdf/src/urdf_sensor_import.cpp
@@ -203,7 +203,7 @@ iDynTree::SensorsList sensorsListFromURDFString(KDL::CoDyCo::UndirectedTree & un
         std::cerr << "Error in loading ft sensors information from URDF file" << std::endl;
     }
 
-    for(int ft_sens = 0; ft_sens < ft_sensors.size(); ft_sens++ )
+    for(size_t ft_sens = 0; ft_sens < ft_sensors.size(); ft_sens++ )
     {
         iDynTree::SixAxisForceTorqueSensor new_sens;
